pyproject.toml
[project]
name = "saxoflow"
version = "1"
description = "Open-source professional RTL design and verification flow"
authors = [{name="SaxoFlow Labs", email="saxoflowlabs@gmail.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "click",
  "questionary",
  "colorlog",
  "packaging",
  "python-dotenv",
  "requests",
  "jinja2",
  "pyyaml",
  "langchain>=0.2",
  "langchain-openai>=0.1.7",
  "langchain_anthropic",
  "langchain_google_genai",
  "rich",
  "prompt_toolkit",
  "pyfiglet"
]

[project.scripts]
saxoflow = "saxoflow.cli:cli" 

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["saxoflow", "saxoflow_agenticai"]